Add guide to 'plugin shading' in Minecraft
A new file 'plugin-shading.md' has been introduced in the minecraft directory.
It gives developers a solution on how to use the shadow plugin to include libraries which are not available on Maven Central when developing for Minecraft. The link for further guidance on shading dependency into a jar file is also attached. This document is added to fill the information gap on how to make a plugin independent from its dependencies when those are not listed on a standard repository.